requirements.txt
Django==1.4
South==0.7.5
amqplib==1.0.2
anyjson==0.3.1
boto==2.5.1
celery==2.5.5
distribute==0.6.24
dj-database-url==0.1.2
django-annoying==0.7.6
django-celery==2.5.5
django-heroku-postgresify==0.2
django-picklefield==0.2.1
django-storages==1.1.4
gunicorn==0.14.2
kombu==2.1.8
psycopg2==2.4.5
python-dateutil==1.5
python-google-places==0.8.0
virtualenv==1.7.1.2
virtualenv-clone==0.2.4
virtualenvwrapper==3.5
wsgiref==0.1.2
